pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "dental-segmentation"
version = "0.1.0"
description = "Dental segmentation using DINOv2/3 backbones and PyTorch Lightning"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
  {name = "Arbiom", email = "your.email@example.com"}
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "lightning>=2.0.0",
    "monai>=1.3.0",
    "timm>=0.9.0",
    "opencv-python",
    "matplotlib",
    "scikit-learn",
    "pycocotools",
    "numpy",
    "pillow",
    "tensorboard"
]

[project.optional-dependencies]
dev = [
    "black",
    "isort",
    "flake8",
    "pytest"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.isort]
profile = "black"
line_length = 88

[tool.black]
line-length = 88
target-version = ['py310']